Refactor with independent MultipartParser
The parses assumes the caller will be a Tora ClientFcgi and is very
optimized for this use case.

It tries to reduce GC pressure by:

 - minimizing the number of allocations until the data is delivered to
   `ModNekoApi`
 - keeping its internal buffer size reasonably small

There is a trade-off, however: previously queued messages will keep
referencing old versions of the internal buffer.  This might be tuned
differently in the future, should it prove to cause issues.